requirements.txt
fastapi
uvicorn
python-dotenv
PyJWT
bcrypt
psycopg2
SQLAlchemy
rich